pyproject.toml
[project]
name = "synthesis-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "chromadb>=1.0.20",
    "docling>=2.73.1",
    "langchain-core>=0.3.76",
    "langchain-openai>=0.3.29",
    "langchain-openrouter>=0.1.0",
    "langdetect>=1.0.9",
    "langgraph>=0.3.34",
    "python-dotenv>=1.1.1",
    "pymupdf>=1.26.4",
    "pydantic>=2.11.0",
    "pdfplumber>=0.11.9",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "jsonschema>=4.25.1",
    "pytest>=8.4.2",
]
